trace-cmd: Use ptrace to trace children of the process
Since there's no good way to know when a child forks, and we
want to update the filters when it does. We must resort to ptrace.

Yes it sucks, but it works.

When performing a trace-cmd record -F -c program

The "-c" switch will cause trace-cmd to ptrace the program. If
the program forks, trace-cmd will update the filters to follow
the children of the program.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>