trace-cmd: Add trace-cmd record --date option
Adding the --date option to trace-cmd record (and extract) will
cause trace-cmd to write a timestamp into the ftrace buffer and
use it to mapped the ftrace buffer's timestamps with gettimeofday.

This will allow reading of the trace.dat file to be able to convert
the timestamps into actual wall time.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>