trace-cmd: Record date first on record, last on extract
Since record resets everything at the beginning, and may even keep
the trace going when done (-k option), it is best to record the date
before recording the trace. This way we can record the date just after
resetting the buffers before starting the trace.

Extract will record all the current data, so it is best to record the date
when the extract is done, and then it is safe to reset the buffers to
capture the date.

We also ignore date with the start option.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>