trace-cmd: Remove duplicate search for ftrace event in blktrace hack
The first try of the blktrace hack, I searched for a ftrace event.
But then I found that TRACE_BLK is not the same on all kernels so I
did a progressive search for a ftrace event to find where TRACE_BLK may
be.

The linear search was left over from the first attempt and no longer
needed.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>