trace-cmd: Add --func-stack option to enable function stack trace
Added --func-stack option to trace-cmd record, that will enable the
function stack tracing, but only if the -l option (limit functions)
is used, and succeeds in filtering functions. It is known that running
the function stack trace on all functions can livelock the machine.

The output now looks like this:

          <idle>-0     [002] 64003.696572: function:             schedule
          <idle>-0     [002] 64003.696577: kernel_stack:         <stack trace>
=> cpu_idle (ffffffff8100a38c)
=> start_secondary (ffffffff814ab828)
       trace-cmd-4132  [003] 64003.696632: function:             schedule
       trace-cmd-4132  [003] 64003.696637: kernel_stack:         <stack trace>

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>