trace-cmd: Enable recording of all CPUs instead of just online ones
Count all possible CPUs for recording instead of online ones.
If we do just online ones, we may not record the ones that are online.
We currently just count the CPUs that are on line. If we take CPU 1
offline leaving 3 CPUs online, we record incorrectly CPUs 0-2 instead
of CPUs 0,2,3.

There may also be a case that we want to record taking a CPU offline
or bringing one back on.

Reported-by: Steve Clark <Steve.Clark@ffei.co.uk>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>