trace-cmd: Update man pages for 'check-events' option
There is a new option added to the 'report' target to check event
formats of the events recorded in a trace data file. This patch updates
the documentation in the man page.

This patch also adds a new man page for 'check-events' target which
verifies event formats on the local machine.

Cc: Michael Rubin <mrubin@google.com>
Cc: David Sharp <dhsharp@google.com>
Signed-off-by: Vaibhav Nagarnaik <vnagarnaik@google.com>
Link: http://lkml.kernel.org/r/1313091459-12049-5-git-send-email-vnagarnaik@google.com
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>