trace-cmd: Do not use threads for extract
Currently the trace-cmd extract function uses the forked threads to
read the data from the ring buffers. As this is not a live trace and
the program does not end till all the buffers are flushed to disk,
there is no need to use the threads. Just serially read each buffer
into a file and then pull them together normally.

This also fixes a bug that was triggered in the kernel where reading
the trace_pipe_raw file after EOF will yield the last page again,
causing the trace-cmd extract to produce duplicate pages.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>