trace-cmd: fix length calculation for padding events
Padding events store size in bytes, not words. Usually this ends up pushing
the parser off the current page, but occasionally not, and when not,
it ends up desynchronizing the parser from the event stream. This would
manifest as a lot of "ug!" messages from trace-cmd, and kernelshark
crashing before displaying the UI.

Link: http://lkml.kernel.org/r/1327633215-4013-1-git-send-email-dhsharp@google.com

Signed-off-by: David Sharp <dhsharp@google.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>