parse-events: Rename parse-events.[ch] to event-parse.[ch]
As the libtraceevent now lives in the kernel repository, but we
can not start using it until it is shipped by all distros, we need to
keep up with the changes that are made there.

This renames the parse-events.* files to event-parse.* as that's what
they are called in the kernel directory.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>