trace-cmd: Don't fail with truncated file for no data
If nothing was traced, the trace-cmd report fails with a truncated file
warning, making it seem that the file was corrupted. But in reality
it is a miscalculation due to the check of the offset and size compared
to the total file size:

   if (offset + size > total_size)
       Must be truncated???

But this is not true. If all CPUs are zero, the offset is always set
to be paged aligned to the last written part of the file. If the cpu
data is empty (size is 0), this alignment may move the offset pass
the end of file, triggering this warning.

The warning should only trigger if size is greater than zero.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>