tools lib traceevent: Make plugin options either string or boolean
When a plugin option is defined, by default it is a boolean (true or false).
If the option is something else, then it needs to set its "value" field to
a default string other than NULL (can be just ""). If the value is not set
then the option is considered boolean, and the updating of the option value
will be handled accordingly.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>