kernelshark: Don't indent and show parents of function events
The indent of functions with the tree view in kernelshark and tree-view
is broken, as it requires the function events to be read in order,
and the GUIs do not do that consistently.

Now that the function plugin has options to disable indent and
enable printing the parent, default to that instead. Now the functions
can be read normaly again.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>