trace-cmd: Speed split up with quicker start time search
Instead of reading record by record for the start time, use
tracecmd_set_cpu_to_timestamp() function. This function does a
bisect of the file to find the page that contains the start
time, then the split operation only needs to search for the
record on that page. This greatly reduces the time to find
a timestamp for very large files.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>