trace-view: Fix missing ref count for saved record
When the record passed in to task_plot_event() doesn't match
the pid of the task, but there exists no record saved before,
then the record passed in is saved as a reference point.

But in this case, the record needs to have its ref count incremented
because the calling function will free it. That was missing and
caused a bug when the record was freed again.

Reported-by: Kaess, Richard <Richard.Kaess@siemens.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>