trace-cmd: Fix size_t to unsigned long long in trace-output.c
When fixing a bug where an 'int' was used, I thought that I should
convert all numbers to size_t. Well, it turns out that size_t
is not 64 bits on 32bit machines. This causes problems when reading
64 bit trace.dat files on 32bit machines.

Add a new 'tsize_t' and 'stsize_t' to trace-output.c where it is
always 64 bits, and should work on both 32 bit and 64 bit machines
while reading any size data file.

Reported-by: Kalle Valo <kvalo@adurom.com>
Tested-by: Kalle Valo <kvalo@adurom.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>