kbuffer: Fix check of missed events
Only the first event on a new page can have missed events before
it. A check is made to make sure that this event is the first
event before checking the page data for missed events.

Unfortunately, the kbuf->index was used, which is the raw index
into the page. It will never be zero, as that would be the location
of the page header itself. kbuf->curr is the index into the data
part of the page. Use that instead.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>