trace-cmd: Allow for updating of option data
Add tracecmd_update_option() and have tracecmd_add_option() return
a descriptor for the option. This way the user can add an option and
then update it later. The update must be smaller or equalt to than
what the original option allocated.

Care must be taken, as the file will still contain the original size
of the option.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>