trace-cmd: Add recording to buffer instances
Add a '-B' option that allows trace-cmd record to created and write
to buffer instances. That is:

  /sys/kernel/debug/tracing/instances/<buffer>

These extra buffers is added to Linux 3.10 and are agnostic from
the main buffer (as well as other instances).

The format is:

  trace-cmd record -p function -e sched -B foo -e block -B bar -e irq:softirq_entry

The above will enable the function tracer and all sched events for the
main buffer, but it will also create two buffer instances called
'foo' and 'bar'. The 'foo' buffer will record all block events where
as the 'bar' buffer will only record the irq:softirq_entry event.
Each of these buffers will be separate. The main buffer is recorded in
the trace.dat file as normal, but a new trace option is added called
TRACECMD_OPTION_BUFFER that stores the offset of where the buffers
cpu data is locate, and the buffer's name.

Next we need to get trace-cmd report to show this data, as it simply
ignores it now.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>