tracing: Add trace-cmd record -D option to force proc use of ftrace_enabled
Now that -d will try to use the function-trace option if it exists, and will
not touch the /proc/sys/kernel/ftrace_enabled file, the -D option is added
to do just that. -D makes trace-cmd work like -d use to, but it will also
enable or disable the function-trace option too (if it exists).

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>