trace-cmd: Add RT specific timings for wakeup info
When running trace-cmd report with -w, the report will look at the times
a task was woken up to the time it was scheduled and report the wakeup
latency. But real time tasks are usually more interesting to know than normal
SCHED_OTHER tasks.

Record real time tasks separately as well and report the real time task
data at the end, along with the normal report.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>