trace-cmd: Add -m option to record to limit size of data file
As long runs of recording traces can cause the disk to fill up, adding
an option that lets the user limit the size of the data file can be really
useful.

Now with the '-m' option to trace-cmd record, this will limit the max size
of the per cpu data files to this number in kilobytes (rounded to pages).

Suggested-by: David Sommerseth <davids@redhat.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>