trace-cmd: Add -M option to set tracing_cpumask to record/start
Add the -M option that takes a parameter to enable the tracing_cpumask.
This works with both record and start.

If the -M option comes after a buffer instance (-B) then it will
only affect that buffer instance cpu mask.

If '-1' is the option, it will set the cpumask to all CPUs.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>