trace-cmd: Fix filter on pids to work more robustly
The filtering on pids with --pid option for trace-cmd report was
quite error prone. It would miss lots of useful events that are
related to the pids. As the filter logic itself is quite robust,
we do not need to specify the events for the fields they may be
in, like prev_pid in sched_switch. We can just use pid and prev_pid
and next_pid, directly and the filter code will figure out what
events it needs.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>