trace-cmd: Have report handle the trace_clock bug
The trace_clock data was placed in a bad location, and it can cause
a file to be corrupted. That bug is fixed, but if a file is corrupted
then have the trace-cmd report still be able to read the content
of the trace. Although the trace clock data may not match.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>