trace-cmd: Use "python2" executable name
Here is my small patch that changes "python" to "python2" in trace-cmd in
some places; it is needed to build and run trace-cmd python parts on
systems where python3 is the main one, for example, Arch Linux. As far as I
know, all distributions have "python2" symlink so this should not break
anything.

Signed-off-by: Nikolay Amiantov <nikoamia@gmail.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>