trace-cmd: Add trace-cmd report --event option
Add an --event option for trace-cmd report that lets the user pass in
a regular expression that will only list the events in the file that
match the regex given.

 trace-cmd report --event ftrace

Will list all ftrace events.

 trace-cmd report --event sys:read

Will list all the events where the system matches "sys" and the event
name matches "read".

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>