trace-cmd: Fix bug returning realloced option as descriptor
The options can be preallocated and used later. But the code had a major
fail in that it returned the address of options from a list that could
be reallocated! That means the handle to the descriptor that was passed
to users could no longer point to the descriptor, and all hell could
break loose.

Convert the options array in the tracecmd_output handle to a link list.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>