trace-cmd: Add -B option to list command to show buffer instances
Add trace-cmd list -B to list the buffer instances that are defined on
the system.

Also cleaned up the code a bit by adding a show_all variable that gets
cleared when any list option is set. This gets rid of the long if statement
that checks to make sure no option was set before deciding to display
everything.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>