trace-cmd: Have -t option reset instance to top_instance
In case of other option affecting the current instance, the -t option
needs to set the current instance to point back to top_instance, otherwise
options specific to the current instance will not affect it.

This is needed for something like:

 trace-cmd reset -B foo -b 1 -t -b 1407

Where foo should have its buffer reset back to 1 and the top buffer be
reset back to 1407, without this fix, foo will get the 1407 and the
top instance will be ignored.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>