trace-cmd/msg: Use poll(2) to wait for a message
Use poll(2) to wait for a message. If a client/server cannot send a message for
any reasons, the current server/client will wait in a blocking read operation.
So, we use poll(2) for avoiding remaining in a blocking state.

Changes in V4: Change the argument of tracecmd_msg_recv_wait()
               Fix some typos

Link: http://lkml.kernel.org/r/20140711005829.25516.75538.stgit@yuno-kbuild.novalocal

Signed-off-by: Yoshihiro YUNOMAE <yoshihiro.yunomae.ez@hitachi.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>