trace-cmd: Add helper to read *_stack ftrace events
The user_stack and kernel_stack events are special in that they are dynamic
arrays but aren't labeled as __data_loc like other dynamic arrays.  This isn't a
big deal since they look the same, they have a "caller" field which is an array
of unsigned longs.  This patch adds a helper to the python bindings to read
these arrays in, lookup the function and append them to a python list.

Link: http://lkml.kernel.org/r/1411667037-27105-1-git-send-email-jbacik@fb.com

Signed-off-by: Josef Bacik <jbacik@fb.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>