trace-cmd: Add --profile to report
Add an option to trace-cmd report called '--profile' that will read the
events from trace-cmd record --profile and produce statistics of where
userspace was in the kernel as well as where tasks were blocked the
longest (shown by stack traces).

This is still a rough version and this will be cleaned up.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>