trace-cmd: Add trace-cmd stream
Add 'trace-cmd stream' which streams data from the ring buffer binary
files and converts them to human readable format and outputs them to
stdout. This is mainly added to be the infrastructure of the profile
command.

Stream is the same as record except that it does not take an output
file parameter (-o), and does not write a file.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>