trace-cmd: Add tracecmd_get_tracing_dir() that does not need to be freed
There were lots of locations that did not free the tracing_dir from
tracecmd_find_tracing_dir(). But that function will allocate the string
and not reference it again. It must be freed. The tracecmd_output that
uses this requires it to be freed.

Create a new tracecmd_get_tracing_dir() for those locations that do
not require it to be freed.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>