parse-events: Copy trace_clock and free it
The pevent->trace_clock should not be a direct pointer to what
was given. It should be copied and freed.

Note, valgrind pointed this out when a caller passed in a pointer
that needed to be freed and it never was. Ideally, pevent should
copy it (which this change does), and free the copy. It's up
to the caller to free the clock string passed in.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>