trace-cmd: Add trace-cmd profile
Add 'trace-cmd profile' that runs like trace-cmd stream does except
that it does not print the events as they arrive, but instead
analyzes the data and creates a profile output at the end.

When stated, it will enable function graph tracing if the depth
option is supported, and will set the depth to 1 (record just
the function that enters and leaves the kernel). The sched events
are enabled with stack tracing, and at the end it will show where
the task entered and exited and for how long. It will also show
where tasks have been blocked.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>