trace-cmd: Remove lost profile start items when events are lost
If events are lost, then that can cause issues with timings of events
as we can have something like:

	sched_switch(old_task, new_task)

	<lost events>
	sched_wakeup(old_task);
	sched_switch(new_task, old_task);
	sched_switch(old_task, new_task);
	</lost_events>

	sched_wakeup(old_task);

and the wakeup will look lik it was much longer than it actually was.
To prevent this, if lost events were detected, then we need to remove
all the start holders from the current CPU, and also ones that can
migrate.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>