trace-cmd: Flush out stdout before forking in record
stdout flushes on new line when writing to console. But when stdout
is redirected to a file, it does a larger buffering when writing.
Unfortuately, this means that the output may still be in the buffer
when we fork. This causes the output for things like the filter files
and version number to be duplicated for each instance we fork.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>