trace-cmd profile: Make the event hash key a bit better
The key is only using the event_data pointer to find the event, but
the search also looks at the search_val and val, which causes lots
of events to fall into the same bucket. Add the search_val and val
fields into the key as well.

Also increase the event buckets to 1024.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>