trace-cmd record: Do not preserve filters, just reset them
If a filter is specified on the command line, after recording, just reset
it instead of trying to put back what was already there. This is because
some filters may be the subsystem filter file, and not the individual
filter file, in which that filter can have issues trying to write back
to it what was found in it.

If the user happens to use a filter, there's no need to put it back when
finished.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>