trace-cmd profile: Do not let delta times be negative
Due to clocks not being in sync across CPUs, recording live events
in the profile that compares the timestamps of different events which
could happen on different CPUs, it could be possible to see time go
backwards.

When this happens, we do not want to record a negative number, but
instead just zero it out.

Ideally, we should switch to the global clock. But better yet, there
should be an option to change the clock.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>