trace-cmd profile: Fix spurious stacktraces from wakeup event
If a wakeup event occurred when a task was not "asleep", which
could happen from dropped events, the wake up is ignored. But if
the proxy task does not have its last-event removed, and if a
stack trace occurs after the wakeup event, it will be hooked to
whatever was the proxy tasks's last event. This caused the output
to show spurious stack traces for various events (mostly the timer
softirq).

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>