trace-cmd extract: Add -a option for all instances
This option will extract all instances that currently exist in the
system, including the top instance. This differs from the meaning of -a
for record and stream (enable all events), which would have no purpose
for extract. Such difference in meaning already exists for -s, so this
seemed reasonable.

Had to fix a bug in create_instance(), in which it ignored its parameter.

Link: http://lkml.kernel.org/r/444ea8355d8c0013ec9d35155d4ea83ec1a7bbab.1429677461.git.hcochran@kernelspring.com

Signed-off-by: Howard Cochran <hcochran@kernelspring.com>
[ Fixed compile error to use different __add_all_instances() ]
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>