trace-cmd: Don't fail build if libaudit is not installed
Add a test compile to make sure that libaudit is installed and define
NO_AUDIT if it is not. When that happens, a warning will occur in building
trace_profile() saying that it will use raw syscall numbers. Doing a
make NO_AUDIT=1, will surpress that warning.

Also fixed the NO_PTRACE to use "override" on changing of CFLAGS as
without it, it has no affect.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>