trace-cmd: Add option to group like comms for profile
When profiling large applications that have multiple children we really want to
see the profile of the app all together.  Add a --by-comm option to accomplish
this for profile.  This merges all the events for like comms into one big chunk
so we can easily see what a giant app is doing, such as HHVM.  Thanks,

Link: http://lkml.kernel.org/r/1432229408-1479-1-git-send-email-jbacik@fb.com

Signed-off-by: Josef Bacik <jbacik@fb.com>
[ fixed up trace_hash ]
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>