trace-cmd/profile: Fix merging of by-comm tasks for schedule
The sched_switch event had the same issue as the sched_wakeup event
in that the pid was used for lookups. Instead of having a separate
merge matching function, use the normal lookup function, but as we
do not care if the pid of the sched_switch or sched_wakeup events
are different, as long as we know that they all belong to the same
comm, we can ignore them. Just need to modify the search vals before
doing the lookups.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>