trace-cmd: Always use /var/run and /var/lib
Looking at documentation on the Internet, it seems that even if a tool is
installed in /usr/local, it still should be using /var/run and /var/lib, as
those are not places that will get overrided by upgrades, but all versions
of trace-cmd should know about it.

I still keep it dynamic for installing in $HOME, as that may not be done by
a administrative user.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>