trace-cmd: Never update saved files when keep is set for recording
If -k is set on the command line, not only do not update the reset_files
list, but also do not call the update of the file even if the reset_files
has something in it. This allows for reset_save_file() to be used during the
parameter parsing before -k may be used.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>