trace-cmd: Use set_event_pid when available
The new set_event_pid file allows the use of filtering pids without needing
to add to event filters directly. By adding a pid to the set_event_pid file,
it will filter all events on the pids listed. It is much more efficient than
the event filter files, and they can handle many more pids than the filter
files can. In fact, this was tested with hackbench, and it was able to still
cope.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>