trace-cmd: Quiet some output when using libtracecmd
If you link against libtracecmd you get a few extraneous warnings and printf's
that aren't related to your tool.  So by default make the path to the events you
are tracing use pr_stat() so if it's not defined we don't get the output, and
make the weak warning() do nothing by default.  The trace-cmd stuff has this
defined so it'll continue to work normally, it just won't be noisy for consumers
of libtracecmd.

Link: http://lkml.kernel.org/r/1449243450-4280-1-git-send-email-jbacik@fb.com

Signed-off-by: Josef Bacik <jbacik@fb.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>