kernelshark: Show event data
Add a display to show an events data in a popup window for the event list.
Add both a normal event output and a raw event.

The normal event is useful when an event has more than one line,
like stack traces, as the rows still don't work quite right.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>