kernelshark: Do not calculate the graph the first time
It appears that showing the window widget calls the configure event for
the graph, which does the calculation of all the events. Then in
gtk_main() that is called again, with a slightly different width
(perhaps the window manager changed it?). This causes the same calculations
to happen again. When it takes 15 seconds (or more) to draw the graph
due to millions of events, doubling the time is quite annoying.

Add a flag to not draw the events the first time through. Then draw them
from then on.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>