kerenshark: Have task plots show when they are in interrupt context
Have the task plots display when they are in interrupt context. If two
events together have their HARDIRQ flags set, then the box that fills them
will be hollow. When the cursor is over that hollow box, it will display in
the info box "(in interrupt)" to let users know why its hollow.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>